Many changes for 1.1.6
Tested in WordPress 4.9
Fixed issue where widget Signup buttons on tournaments with one or more
participants are not displayed for users not logged in. Thanks to
sagund07 for finding the cause and giving a solution for this issue.
Added comments in code about relevant functionality changes in WordPress
4.6.
Changed some class names in settings to be less conflicting with other
plugin CSS that may be loaded as well.
Added a way to list participants to shortcodes.
Fixed insecure content loading in iframes on secure pages.
Improved caching for much faster page loads.
Added adaptive caching where the expire date is determined by the cached
content.
Added cache age to bottom corner shortcodes and widgets which also
allows manual refreshing.